update: patched Dockerfile
This commit updates the dockerfile to use multistage build for the Node.js application. This removes the requirement to run npm install locally before building the image. The first stage installs the dependencies and builds the application, while the second stage copies only the necessary files and runs the application. The dockerfile also uses alpine as the base image for both stages, resulting in a smaller and more secure image.